[Fix] Made Skills Table code into MVC
- Created skillTable component (SkillTable & SkillTableView)
